azureml_sample/requirements.txt
azure-ai-ml
azure-identity
pandas
requests
pyarrow
scikit-learn
joblib
